Refactor rounds into own files in participant demo (#20)
